pyproject.toml
[project]
name = "the-main-quest"
version = "0.1.0"
description = ""
authors = [
    {name = "kolikaran"}
]
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "dynaconf (>=3.2.12,<4.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "python-json-logger (>=4.0.0,<5.0.0)",
    "requests (>=2.32,<3.0)",
    "psycopg2-binary (>=2.9,<3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "ipykernel (>=7.2.0,<8.0.0)",
    "ipywidgets (>=8.1.8,<9.0.0)"
]
